refactor:[#S05P21B304-1053] Refresh Token
Former-commit-id: 6272a1e36f4e9f826bf256f677576c1ee296ce27

+++ back-end/AniChew/src/main/java/com/anichew/Service/UserServiceImpl.java
@@ -225,9 +225,11 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public void deleteRefreshToken(HttpServletRequest httpServletReq, HttpServletResponse httpServletRes) {
 		
-		String refreshToken = cookieUtil.getCookie(httpServletReq, jwtUtil.REFRESH_TOKEN_NAME).getValue();
-		
-		boolean deleted = redisUtil.deleteData(refreshToken);
+		Cookie refreshTokenCookie = cookieUtil.getCookie(httpServletReq, jwtUtil.REFRESH_TOKEN_NAME);
+		if(refreshTokenCookie != null) {
+			String refreshToken = refreshTokenCookie.getValue(); 		
+			boolean deleted = redisUtil.deleteData(refreshToken);
+		}
 		cookieUtil.deleteCookie(httpServletReq, httpServletRes, jwtUtil.REFRESH_TOKEN_NAME);
 		
 	}
